refactor: consolidate type definitions for single source of truth
- Consolidate AgentSessionDescriptor: Move schema to AgentDataclass.ts
  and derive type from it using z.infer<>. AgentSessionSchema.ts now
  re-exports for backward compatibility.

- Unify usage types: Create UsageTypes.ts as single source of truth
  for TokenUsageStats, ExtendedTokenUsageStats, ResponseUsageBase,
  provider-specific usage types, and accumulation types. Update
  ResponseUsage.ts and RunUsageAccumulator.ts to import from there.

- Rename ToolUseSnapshotCache to ToolUseSessionManager: The canonical
  class name is now ToolUseSessionManager. ToolUseSnapshotCache is
  exported as a deprecated alias for backward compatibility.

- Decouple tools from agent core: Create ToolTypes.ts with ITool,
  IToolRegistry, and ToolResult interfaces in @agent/core. Tools now
  implement ITool interface. BaseToolUseAgent accepts optional
  toolRegistry via constructor options, defaulting to
  DEFAULT_TOOL_REGISTRY. This enables dependency injection for testing.
