+++ src/agent/core/AgentSessionSchema.ts
@@ -1,13 +1,5 @@
-// Third-party imports
-import { z } from 'zod';
-
-// Local imports - agent
-import { AgentCategory, AgentType } from './AgentDataclass';
-
 /**
- * Canonical schema for agent session descriptors shared across agent modules.
+ * @deprecated Import AgentSessionDescriptorSchema from './AgentDataclass' instead.
+ * This file re-exports for backward compatibility.
  */
-export const AgentSessionDescriptorSchema = z.strictObject({
-  agentType: z.enum(AgentType).optional(),
-  agentCategory: z.enum(AgentCategory),
-});
+export { AgentSessionDescriptorSchema } from './AgentDataclass';
