+++ src/agent/core/index.ts
@@ -4,8 +4,11 @@ export * from './AgentState';
 export * from './AgentSharedStore';
 export * from './AgentSharedStoreRegistry';
 export * from './ToolConfig';
+export * from './ToolTypes';
 export * from './AgentWorkspaceState';
+export * from './UsageTypes';
 export * from './ResponseUsage';
+export * from './RunUsageAccumulator';
 export * from './IAgent';
 export * from './ResponseCycle';
 export * from './ToolUseCycle';
