+++ src/agent/toolUse/ToolUseSessionManager.ts
@@ -1,4 +1,9 @@
-export { ToolUseSnapshotCache as ToolUseSessionManager } from './ToolUseSnapshotCache';
+/**
+ * @file ToolUseSessionManager.ts
+ *
+ * Public exports for tool-use session management.
+ */
+export { ToolUseSessionManager } from './ToolUseSnapshotCache';
 export type {
   ToolUseSessionSnapshot,
   SaveToolUseSnapshotPayload,
