+++ src/agent/types/UsageTypes.ts
@@ -1,45 +1,15 @@
 /**
- * Token usage statistics for tracking model usage and costs.
+ * @file UsageTypes.ts
+ *
+ * Re-exports usage types from the consolidated source in @agent/core/UsageTypes.
+ * This file is kept for backward compatibility with existing imports.
+ *
+ * @see @agent/core/UsageTypes for the single source of truth.
  */
-// Third-party imports
-import { z } from 'zod';
-
-export const TokenUsageStatsSchema = z.strictObject({
-  /** Number of input tokens consumed */
-  inputTokens: z.number(),
-  /** Number of output tokens generated */
-  outputTokens: z.number(),
-  /** Total cost in USD for the request */
-  cost: z.number(),
-});
-
-export type TokenUsageStats = z.infer<typeof TokenUsageStatsSchema>;
-
-/**
- * Extended statistics tracked during agent execution.
- */
-export interface ExtendedTokenUsageStats extends TokenUsageStats {
-  /** Total elapsed time in seconds */
-  elapsedTime?: number;
-  /** Tokens read from cache */
-  cacheReadInputTokens?: number;
-  /** Tokens written to cache */
-  cacheCreationInputTokens?: number;
-  /** Percentage of tokens served from cache */
-  percentageCached?: number;
-  /** Tokens used for reasoning */
-  reasoningTokens?: number;
-  /** Tokens consumed by tool use */
-  toolUseTokens?: number;
-}
-
-/**
- * Message interface for updating usage stats in the progress view.
- */
-export const StreamUsageMessageSchema = z.strictObject({
-  command: z.literal('updateUsage'),
-  stream: z.string(),
-  usageByRun: z.record(z.string(), TokenUsageStatsSchema).prefault({}),
-});
-
-export type StreamUsageMessage = z.infer<typeof StreamUsageMessageSchema>;
+export {
+  ExtendedTokenUsageStats,
+  StreamUsageMessage,
+  StreamUsageMessageSchema,
+  TokenUsageStats,
+  TokenUsageStatsSchema,
+} from '@agent/core/UsageTypes';
